Correctif sur la suppression du dossier extrait de l'archive zip.
git-svn-id: http://svn.sylogix.org/gepiphp/branches/release-1.5.5@8292 e073496e-d129-0410-8f7b-841bfd467322

+++ mod_trombinoscopes/trombinoscopes_admin.php
@@ -262,8 +262,20 @@ if (isset($_POST['action']) and ($_POST['action'] == 'upload'))  {
 					}
 					if($nb_photos_eleves>0) {$msg.="$nb_photos_eleves photo(s) élève(s) traitée(s).<br />";}
 					closedir($dossier);
-					if (!rmdir ($folder)) {
-						$msg .= "Erreur lors de la suppression de ".$folder;
+					if(file_exists($folder)) {
+						$dossier = opendir($folder);
+						while ($Fichier = readdir($dossier)) {
+							if ($Fichier != "." && $Fichier != "..") {
+								$source=$folder."/".$Fichier;
+								if (!@unlink ($source)) {
+									$msg .= "Erreur lors de la suppression de ".$source;
+								}
+							}
+						}
+						closedir($dossier);
+						if (!rmdir ($folder)) {
+								$msg .= "Erreur lors de la suppression de ".$folder;
+						}
 					}
 				}
 	
@@ -297,8 +309,20 @@ if (isset($_POST['action']) and ($_POST['action'] == 'upload'))  {
 					}
 					if($nb_photos_personnels>0) {$msg.="$nb_photos_personnels photo(s) personnel(s) traitée(s).<br />";}
 					closedir($dossier);
-					if (!rmdir ($folder)) {
-						$msg .= "Erreur lors de la suppression de ".$folder;
+					if(file_exists($folder)) {
+						$dossier = opendir($folder);
+						while ($Fichier = readdir($dossier)) {
+							if ($Fichier != "." && $Fichier != "..") {
+								$source=$folder."/".$Fichier;
+								if (!@unlink ($source)) {
+									$msg .= "Erreur lors de la suppression de ".$source;
+								}
+							}
+						}
+						closedir($dossier);
+						if (!rmdir ($folder)) {
+								$msg .= "Erreur lors de la suppression de ".$folder;
+						}
 					}
 				}
 			}
